% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{show,glpModel-method}
\alias{show,glpModel-method}
\alias{show,glpModel,ANY-method}
\title{Show method for \code{glpModel}}
\usage{
\S4method{show}{glpModel}(object)
}
\arguments{
\item{object}{An object of class \code{glpModel}.}
}
\description{
Show method for \code{glpModel}
}
